5.6.31 GUI & Maintenance
* Security update : Encryption of the panel file in the macOS bundle for restricted instances. CtrlrMac.cpp, CtrlrEditorApplicationCommandsMenu.cpp, CtrlrManager.cpp

* NEW JUCE Class MouseInputSource added to LUA. LCore.cpp,LCore.h, LJuce.cpp, LJuce.h, LMouseInputSource.h
* NEW algorithm for Roland checksums. CtrlrSysexProcessor.cpp & CtrlrUtilities.cpp
* NEW settings for LUA Method Editor. CtrlrValueTreeEditor.h CtrlrLuaMethodCodeEditorSettings.cpp & .h CtrlrLuaMethodCodeEditor.cpp CtrlrLuaMethodEditor.cpp CtrlrIDs.h & CtrlrIDs.xml

* FIXED MSB 14 bit numbers sending 0xFF instead of 0x7F. CtrlrSysexProcessor.cpp & CtrlrUtilities.cpp
* FIXED missing File Management bottom notification bar. CtrlrPanelFileOperations.cpp, CtrlrPanelEditor.cpp, CtrlrPanelEditor.h, CtrlrPanel.cpp, CtrlrPanel.h
* FIXED missing menuBar on export for log and MIDI monitor windows. CtrlrChildWindowContainer.cpp
* FIXED uiSlider value not reaching maxValue when using negative values for minValue. CtrlrSlider.cpp
* FIXED uiImageSlider value not reaching maxValue when using negative values for minValue. CtrlrImageSlider.cpp
* FIXED black text on black background for Modulator List window. CtrlrPanelModulatorList.cpp
* FIXED typo in alert when closing dirty panel. CtrlrPanelFileOperations.cpp
* FIXED parameters count passed to VST host is set from the highest vstindex when a panel is an exported VST/VST3 instance, not (64). CtrlrProcessor.cpp
* FIXED LnF panel close button colour on mouseover follows the panel colourScheme. CtrlrDocumentPanel.cpp
* FIXED CtrlrModulator Value statement precised to help avoid feedback loops between LUA and (delayed) UI. Commit 6e5a0b2 by midibox. CtrlrLuaManager.cpp
* FIXED exported VST crashing DAW if panelIsDirty = 0 on export. CtrlPanelFileOperations.cpp
* FIXED Ctrlr not showing up in Ableton Live. CtrlrX.jucer, CtrlrProcessor.cpp & CtrlrProcessor.h, CtrlrProcessorEditorForLive.cpp & CtrlrProcessorEditorForLive.h
* FIXED MIDI Monitor IN/OUT turned ON by default. CtrlrManager.cpp
* FIXED Console window & Midi Monitor crashing Cubase if closed from the Menu File>Close. CtrlrLuaConsole.cpp & CtrlrMIDIMonitor.cpp
* FIXED LUA mod:getMidiMessage():getProperty("propertyName") CtrlrMidiMessage.cpp & CtrlrMidiMessage.h
* FIXED exported instances not getting the proper LnF version or colourScheme (popup, child windows etc). CtrlrManagerInstance.cpp, CtrlrEditor.cpp
* FIXED property pane tabs not showing up on exported instances. CtrlrManagerInstance.cpp
* FIXED Property Pane tabs not following panel LnF. CtrlrPanelProperties.cpp & CtrlrPanelProperties.h
* FIXED Child Windows (LUA Editor, console etc) not getting the proper menuBar background. CtrlrChildWindowContainer.cpp
* FIXED LUA Editor Method Tree selected item not getting the proper colour. CtrlrLuaMethodEditor.cpp
* FIXED AU AudioUnit version of CtrlX was failing the Apple/Logic Validation Test. CtrlrX.jucer, CtrlrProcessor.cpp & CtrlrProcessor.h
* FIXED Panel Tabs showing up on exported instances. CtrlrDocumentPanel.cpp
* FIXED messy preferences window and settings. CtrlrSettings.cpp & CtrlrSettings.cpp ctrlrIDs.h & CtrlrIDs.xml
* FIXED default New Panel LnF not following global LnF from Global Preferences. ctrlrPanel.cpp ctrlrIDs.h & CtrlrIDs.xml
* FIXED CtrlrX.ico too dark to be legible on dark backgrounds. ctrlr_logo_circle_v3.svg
* FIXED FileChooser hanging when exporting instance on OSX Catalina & macOS BigSur. CtrlrManager.cpp
* FIXED About popup design refurbished with new CtrlrX logo. CtrlrAbout.cpp & CtrlrAbout.h
* FIXED About popup not getting current build date, fixed with C++ Macro timestamp. CtrlrRevision.h

+++ Source/UIComponents/CtrlrComponents/Sliders/CtrlrImageSlider.cpp
@@ -172,7 +172,7 @@ void CtrlrImageSlider::valueTreePropertyChanged (ValueTree &treeWhosePropertyHas
     }
     else if (property == Ids::uiSliderInterval || property == Ids::uiSliderMax || property == Ids::uiSliderMin)
     {
-        ctrlrSlider->setRange ( getProperty (Ids::uiSliderMin), getProperty (Ids::uiSliderMax), getProperty (Ids::uiSliderInterval) );
+        ctrlrSlider->setRange ( (float) getProperty (Ids::uiSliderMin), (float) getProperty (Ids::uiSliderMax), (float) getProperty (Ids::uiSliderInterval) );  // v5.6.31 (float) instead of plain values. Plain values and (double) were giving false values when negative.
         owner.setProperty (Ids::modulatorMax, ctrlrSlider->getMaximum());
         owner.setProperty (Ids::modulatorMin, ctrlrSlider->getMinimum());
     }
